Changed "unsigned long" to "uint32_t" where 32-bit unsigned int required.
git-svn-id: svn+ssh://svn.nta-monitor.com/trunk/opensource/ike-scan@2448 062a1500-4a13-0410-a63b-ee65f32af78f

+++ sha1.h
@@ -9,12 +9,12 @@
 #define SHA1_H
 
 typedef struct {
-    unsigned long state[5];
-    unsigned long count[2];
+    uint32_t state[5];
+    uint32_t count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 
-void SHA1Transform(unsigned long state[5], unsigned char buffer[64]);
+void SHA1Transform(uint32_t state[5], unsigned char buffer[64]);
 void SHA1Init(SHA1_CTX* context);
 void SHA1Update(SHA1_CTX* context, unsigned char* data, unsigned int len);
 void SHA1Final(unsigned char digest[20], SHA1_CTX* context);
